Removes sample from FoG's Dialogue Module, as we already made our implementation. Also does some small refactorings.
