--- Assets/Samples.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e41b6b3a2f66664caef2480d04d96c2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
