--- Assets/Samples/[FoG] Dialogue Module.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 30da8ab92451e664fb3a242ef1b38f26
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
