--- Assets/Samples/[FoG] Dialogue Module/1.0.0.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5571b908330838246a88ab1cc8f66c9d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
