--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1cda0ec91672d14292510d412128539
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
