--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/InputSystem.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cfe8ca593cc8bc84981b7918baa8274b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
