--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/InputSystem/Inputs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b249d66ceb83bd3498fadb7ffcb7058a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
