--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/InputSystem/Inputs.inputactions.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 2f3a88ff43d40b4408568828b9b40e6b
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
-  generateWrapperCode: 1
-  wrapperCodePath: 
-  wrapperClassName: 
-  wrapperCodeNamespace: Fog.Dialogue.Samples
