--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fb2bb17c0df2ad34abb690ce7e56ca27
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
