--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8b46e921381b0bd4eb13cb26f48b642e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
