--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 632ea13d535931b4a980df13c1adb6a7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
