--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Choice.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 12f143a215ec3a54b828f0207fb41a7e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
