--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/ChooseA.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7822ad54727c52146878e8332ac5a369
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
