--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/ChooseB.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 88558fdd662dc794a913c9edaaf2456d, type: 3}
-  m_Name: ChooseB
-  m_EditorClassIdentifier: 
-  lines:
-    array:
-    - speaker: {fileID: 11400000, guid: 975b2c08788ba4543874afd4c71eacad, type: 2}
-      text: You chose B
