--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/ChooseB.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90a40f2ca2871234a83fec26468dcfea
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
