--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Hello World.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1d232c32f53d7dd4f849799b3b12f4d4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
