--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Lorem Ipsum.asset
@@ -1,25 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 88558fdd662dc794a913c9edaaf2456d, type: 3}
-  m_Name: Lorem Ipsum
-  m_EditorClassIdentifier: 
-  lines:
-    array:
-    - speaker: {fileID: 11400000, guid: 975b2c08788ba4543874afd4c71eacad, type: 2}
-      text: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
-        quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-        Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore
-        eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident,
-        sunt in culpa qui officia deserunt mollit anim id est laborum asdasdasdasdasdasdasd
-        asda fsgfd sd gsdwe tgbsfh sh sb sdgfb adfv afgsghaeg asdf vsvc bs  dfgaw
-        taer gadgsd g\sGADFGSHB FGHSDGFH DF NBDF A A
