--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Lorem Ipsum.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7cb8180cc79131f478dce87a51646dde
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
