--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Speaker.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f56badfc58369ee4ab9026bb7cacaf11, type: 3}
-  m_Name: Speaker
-  m_EditorClassIdentifier: 
-  dialogueColor: {r: 0.627451, g: 0.9405171, b: 1, a: 1}
-  dialogueName: Speaker
-  dialoguePortrait: {fileID: 21300000, guid: 499da02e621f7c94b8ddd034c3fd118d, type: 3}
