--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/ScriptableObjects/Speaker.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 975b2c08788ba4543874afd4c71eacad
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
