--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fe40d16008bd0604c8c2bcb42f70c434
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
