--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/DialogueBlockMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 88558fdd662dc794a913c9edaaf2456d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
