--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/InputActivator.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-using UnityEngine.InputSystem;
-
-namespace Fog.Dialogue.Samples {
-    public class InputActivator : MonoBehaviour {
-        [SerializeField] private InputActionAsset actions;
-
-        private void OnEnable() {
-            actions.Enable();
-        }
-
-        private void OnDisable() {
-            actions.Disable();
-        }
-    }
-}
