--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/InputActivator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 525a9f4b713d39e43b090bceef635f07
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
