--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/OptionsDialogueBlockMovement.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Fog.Dialogue.Samples {
-    [CreateAssetMenu(fileName = "NewOptionsDialogue", menuName = "FoG/DialogueModule/Sample/OptionsDialogueSample")]
-    public class OptionsDialogueBlockMovement : OptionsDialogue {
-        public override void BeforeDialogue(){
-            SimpleMove.instance.BlockMovement();
-
-            base.BeforeDialogue();
-            DialogueHandler.instance.OnDialogueStart -= BeforeDialogue;
-        }
-    }
-}
