--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/OptionsDialogueBlockMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 529705b8796ed924e955d1704552ec7a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
