--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Scripts/SimpleMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2560bdef1a9905a4795d34c8af071bdd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
