--- Assets/Samples/[FoG] Dialogue Module/1.0.0/Example/Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9ae197cf4bcfbcc43b81e5d58fed2fd0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
