+++ Assets/Scripts/Game/NPCs/NpcController.cs
@@ -1,34 +1,27 @@
-﻿using System;
+﻿using System.Linq;
 using Fog.Dialogue;
-using ScriptableObjects;
 using UnityEngine;
 
 namespace Game.NPCs
 {
     public class NpcController : MonoBehaviour, IInteractable {
-        [SerializeField] private Dialogue dialogue = null;
+        [SerializeField] private Dialogue dialogue;
 
         public void Reset() {
-            int nColliders = GetComponents<Collider2D>().Length;
-            // Se so tem um collider, se certifica que ele seja trigger
+            var nColliders = GetComponents<Collider2D>().Length;
             if (nColliders == 1) {
                 GetComponent<Collider2D>().isTrigger = true;
             } else if (nColliders > 0) {
-                bool hasTrigger = HasAtLeastOneTrigger();
-                // Se nenhum deles for, se certifica de que o primeiro deles seja trigger
+                var hasTrigger = HasAtLeastOneTrigger();
                 if (!hasTrigger) {
                     GetComponent<Collider2D>().isTrigger = true;
                 }
             }
         }
 
-        private bool HasAtLeastOneTrigger() {
-            foreach (Collider2D col in GetComponents<Collider2D>()) {
-                if(col.isTrigger) {
-                    return true;
-                }
-            }
-            return false;
+        private bool HasAtLeastOneTrigger()
+        {
+            return GetComponents<Collider2D>().Any(col => col.isTrigger);
         }
 
         public void OnInteractAttempt() {
@@ -38,14 +31,14 @@ namespace Game.NPCs
         }
 
         public void OnTriggerEnter2D(Collider2D col) {
-            Agent agent = col.GetComponent<Agent>();
+            var agent = col.GetComponent<Agent>();
             if (agent) {
                 agent.collidingInteractables.Add(this);
             }
         }
 
         public void OnTriggerExit2D(Collider2D col) {
-            Agent agent = col.GetComponent<Agent>();
+            var agent = col.GetComponent<Agent>();
             if (agent) {
                 agent.collidingInteractables.Remove(this);
             }
